inserted Ability to open & close cart
Functionality to open and close the cart by clicking on the cart button.
It uses "useState" to store the state and render the cart components.
Initially the state is set to false and it can be changed to true when
the cart button is clicked, this causes the cart to be displayed and
then when the "close" button on the cart is clicked then the state is set
to "false" resulting the cart to be removed from the screen.