Order button to show when there are items in cart
A new const has been added to store either a "true" or "false" that
depends on the length of the array containing item. This is then used
inside the jsx code that render the "order" button. When "true" the
button is shown and when "false" the button is not visible.