Add new items of the same type
In the "cartreducer" fucntion at the moment items are added as new items in the
array "updatedItems", but this is incorrect. To correct that a new const
"existingCartItemIndex" has been added which can have an idex or null by
checking old values with new added in the array. Also there is IF/ELSE check
which tests and then sets the values before returning them.