Reducing quantity and deleting item from the cart
Firstly, in the reducer fucntion an IF check is done to see the type of
action required. In this scenario the action type is "REMOVE" which
allow to remove one item at a time depending on the ID. Also there is a
filter method that gets excecuted when "existingitem === 1" for that specific type of meal otherwise only the price and quantity is updated.