Animating the cart when adding items
For making the cart to animate "useState" and "useEffect" have been
utilized. The const called "btnIsHighLighted" store either "true" or
"false", this then used switch different style classes. When the state
is "true" "classes.bump" is assigned otherwise empty string. In addition
to that there is "useEffect" to change the class back. The "useEffect"
incorporates a timer which is resposnible for changing the class back
after 300ms. Finally there is a funtion which is used to reset the timer.